Remove TODO on duplicated work unit wrapper + add explanatory comment
This one comes from me trying to find a good alternative to fix a TODO
in `rivertest/worker.go` over the duplicated work unit wrapper and
factory types. Unfortunately I found that there is no good fix available
because anything we do introduces a cyclic dependency. Having put very
low level types like `Job`, `JobArgs, and `Worker` into the top-level
`river` package unfortunately acts to hugely limit our ability to move
things around.

So instead, I just remove the TODO here since there's nothing to be
done, and replace it with an explanatory comment saying how this isn't
great, but it's the only viable method that we know of.
